+++ tests/test_amr.py
@@ -1,10 +1,12 @@
 #!/usr/bin/env python
-""" kick the tires on the amr machinery """
+"""Kick the tires on the amr machinery"""
 
 from __future__ import print_function
 import argparse
+import sys
+import os
 
-from viscid_test_common import sample_dir, next_plot_fname
+from viscid_test_common import next_plot_fname
 
 import viscid
 from viscid import vutil
@@ -12,7 +14,7 @@ from viscid.plot import mpl
 
 
 def run_test(show=False):
-    f = viscid.load_file(sample_dir + "/amr.xdmf")
+    f = viscid.load_file(os.path.join(viscid.sample_dir, "amr.xdmf"))
     plot_kwargs = dict(patchec='y')
     mpl.plot(f['f'], "z=0.0f", **plot_kwargs)
 
@@ -20,15 +22,17 @@ def run_test(show=False):
     if show:
         mpl.show()
 
-def main():
-    parser = argparse.ArgumentParser(description="Test calc")
+def _main():
+    parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument("--show", "--plot", action="store_true")
     args = vutil.common_argparse(parser)
 
     run_test(show=args.show)
 
+    return 0
+
 if __name__ == "__main__":
-    main()
+    sys.exit(_main())
 
 ##
 ## EOF