+++ tests/test_hdf5.py
@@ -1,21 +1,23 @@
 #!/usr/bin/env python
-""" test loading a numpy npz binary file """
+"""Test saving fields to an HDF5 + XDMF pair"""
 
 from __future__ import print_function
-import os
 import argparse
+import os
+import sys
 
 import numpy as np
 
-from viscid_test_common import sample_dir, next_plot_fname
+from viscid_test_common import next_plot_fname
 
 import viscid
 from viscid import vutil
 from viscid.plot import mpl
 from viscid.plot.mpl import plt
 
-def main():
-    parser = argparse.ArgumentParser(description="Test xdmf")
+
+def _main():
+    parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument("--show", "--plot", action="store_true")
     parser.add_argument("--keep", action="store_true")
     args = vutil.common_argparse(parser)
@@ -36,7 +38,7 @@ def main():
     b['z'] = -Zcc
 
     # save an hdf5 file with companion xdmf file
-    h5_fname = sample_dir + "/test.h5"
+    h5_fname = os.path.join(viscid.sample_dir, "test.h5")
     viscid.save_fields(h5_fname, [psi, b])
 
     # load the companion xdmf file
@@ -57,8 +59,10 @@ def main():
         os.remove(h5_fname)
         os.remove(xdmf_fname)
 
+    return 0
+
 if __name__ == "__main__":
-    main()
+    sys.exit(_main())
 
 ##
 ## EOF