+++ viscid/plot/mvi.py
@@ -426,6 +426,10 @@ def points3d(*args, **kwargs):
             `mayavi.mlab.points3d`
 
     Keyword Arguments:
+        modify_args (bool): if True (default), then check if args is a
+            single 2d sequence of shape 3xN or Nx3. Then split them up
+            appropriately. if False, then args are passed through
+            to mlab.points3d unchanged, nomatter what.
         cmap (str, None, False): see :py:func:`apply_cmap`
         alpha (number, sequence): see :py:func:`apply_cmap`
         clim (sequence): see :py:func:`apply_cmap`
@@ -435,7 +439,20 @@ def points3d(*args, **kwargs):
     Returns:
         TYPE: Description
     """
+    modify_args = kwargs.pop('modify_args', True)
     kwargs, cmap_kwargs = _extract_cmap_kwargs(kwargs)
+
+    if modify_args and len(args) < 3:
+        a0 = np.asarray(args[0])
+        if len(a0.shape) > 1 and a0.shape[0] == 3:
+            args = [a0[0, :].reshape(-1),
+                    a0[1, :].reshape(-1),
+                    a0[2, :].reshape(-1)] + list(args[1:])
+        elif len(a0.shape) > 1 and a0.shape[1] == 3:
+            args = [a0[:, 0].reshape(-1),
+                    a0[:, 1].reshape(-1),
+                    a0[:, 2].reshape(-1)] + list(args[1:])
+
     points = mlab.points3d(*args, **kwargs)
     apply_cmap(points, **cmap_kwargs)
     return points